cambios en los filtros y se agregan tests de chofer

+++ src/empresa_naviera/BuqueViaje.java
@@ -59,7 +59,7 @@ public class BuqueViaje {
 	 * @return
 	 */
 	 
-	public double getDuracion() {
+	public int getDuracion() {
 		
 		return circuito.getDuracion();
